fix(watermark): update default watermark to -1
This change is to fix issue #133 where watermark shows -62135596800 when a fresh pipeline started without any traffic. The daemon service shows -62135596800 because when the OT store is empty, the watermark fetcher returns time.Time{} as default watermark, which transforms to epoch as -62135596800. In this change I am proposing -1 as default watermark to indicate there is no valid watermark yet.

Signed-off-by: Keran Yang <yangkr920208@gmail.com>