refactor: use exact matching instead of regex to perform e2e data validation. (#443)
In this PR we switch from regex matching to exact matching. By doing so we avoid using redis.KEYS api which is scanning the entire store to find matches, hence improve the query performance.

The new approach uses pipelineName:sinkName as hash key. Under each hash key, we have the exact data as field and no. of occurrences of the data as value.

Signed-off-by: Keran Yang <yangkr920208@gmail.com>