chore: use dateparse for extracting event time from payload (#597)
This change uses https://github.com/araddon/dateparse for parsing event time string to `time.Time` object. The benefit of using this library is that it can parse many date strings without knowing format in advance, which makes our job easier when we want to parse epoch time string, because we don't need to implement parsing logics for different granularities.

This change also marks the `format` specification as optional. If specified, we will continue using time.Parse to parse the data. But if user doesn't specify `format`, it's also fine as we use `dateparse` to find the format based on time string.

The only uncertainty is that we haven't tested the performance of `dateparse` under a high-traffic numaflow environment. Based on https://github.com/araddon/dateparse/blob/master/bench_test.go, it seems performing pretty well. If we ever find `dateparse` being the bottleneck of performance, we should fallback and implement our own approach to parse epoch.

Signed-off-by: Keran Yang <yangkr920208@gmail.com>