chore: fix TestBuiltinEventTimeExtractor e2e test
With idle watermark in place, the watermark propagation becomes faster than before, that it seems the edge watermark can be the event time of the latest message. Since the purpose of this specific test case is to verify the event time is extracted from the message payload, I am updating the test to pass if the watermark is the event time extracted from the latest message.

Signed-off-by: Keran Yang <yangkr920208@gmail.com>