refactor: avoid exposing internal data structures of pod tracker to the rater (#902)
This is a small refactor to modularize the pod tracker. The data structure of UniqueStringList is the implementation detail of pod tracker, hence should not be exposed to the rater. This makes less work for the rater as rater doesn't need to call getActivePods followed by .Contains to check if a pod is active. This also removes the risk of the rater directly manipulating the UniqueStringList. Lastly, if in the future, we decide to replace UniqueStringList with another data structure, no code change needs to be made on the rater side.

Signed-off-by: Keran Yang <yangkr920208@gmail.com>